feat: query-raw and contract-info command added to wasm

+++ packages/cli/src/config/index.ts
@@ -183,15 +183,11 @@ export async function loadDefaultEnv() {
 /**
  * Loads the config used by the CLI on startup
  */
-export async function loadEnv(env: string, data?: Partial<ReturnType<typeof envConfig.getProperties>>) {
+export async function loadEnv(env: string) {
   const all_envs = Array.from(new Set(envConfig.get('envs').concat(env)));
-  try {
-    const parsedEnvConfig = loadStorageFile(env, 'env.json');
-    envConfig.load(JSON.parse(parsedEnvConfig.toString()));
-  } catch (error) {
-    console.log("\nEnv not found")
-    createEnv(env, data)
-  }
+  const parsedEnvConfig = loadStorageFile(env, 'env.json');
+  envConfig.load(JSON.parse(parsedEnvConfig.toString()));
+
   envConfig.set('envs', all_envs)
   setGQLSdkUri(envConfig.get('gql'));
   await displaySpinnerAsync("Loading config...", loadDefaultConfig);
@@ -201,8 +197,11 @@ export async function loadEnv(env: string, data?: Partial<ReturnType<typeof envC
 /**
  * Loads the config used by the CLI on startup
  */
-export function createEnv(env: string, data?: Partial<ReturnType<typeof envConfig.getProperties>>) {
+export function createEnv(env: string, data?: Partial<ReturnType<typeof envConfig.getProperties>>, doNotThrow = false) {
   if (fs.existsSync(path.join(CONFIG_DIRECTORY, env, 'env.json'))) {
+    if (doNotThrow) {
+      return;
+    }
     throw new Error(`Env - ${env} already exists`);
   }
   console.log("\nCreating new env")
@@ -218,7 +217,7 @@ export function createEnv(env: string, data?: Partial<ReturnType<typeof envConfi
  */
 export function renameEnv(env: string, newName: string) {
   fs.renameSync(path.join(CONFIG_DIRECTORY, env), path.join(CONFIG_DIRECTORY, newName));
-  envConfig.set('envs', envConfig.get('envs').filter(name => name !== newName).concat(newName));
+  envConfig.set('envs', envConfig.get('envs').filter(name => name !== env).concat(newName));
 }
 
 
@@ -255,7 +254,7 @@ export function loadLocalChains() {
 }
 
 /**
- * Migrates previous version of cli config files
+ * Migrates previous version of cli config files - For dev - Only remove legacy part of this function.
  * @deprecated will be removed in next update
  */
 export function migrateLegacyEnv() {
@@ -263,29 +262,35 @@ export function migrateLegacyEnv() {
     if (fs.existsSync(path.join(CONFIG_DIRECTORY, 'env.json'))) {
       return;
     }
+
+    let defaultEnv: string = DEFAULT_ENVS.TESTNET;
     createEnv(DEFAULT_ENVS.TESTNET, {
       'gql': GQL_URLS.TESTNET,
-    });
+    }, true);
     createEnv(DEFAULT_ENVS.MAINNET, {
       'gql': GQL_URLS.MAINNET,
-    });
+    }, true);
     createEnv(DEFAULT_ENVS.DEVNET, {
       'gql': GQL_URLS.DEVNET,
-    });
+    }, true);
 
-    fs.mkdirSync(path.join(CONFIG_DIRECTORY, 'legacy'), { recursive: true });
-    const filesToMove = ['keys.json', 'config.json', 'chainConfigs.json'];
-    filesToMove.forEach(f => {
-      if (fs.existsSync(path.join(CONFIG_DIRECTORY, f))) {
-        fs.copyFileSync(path.join(CONFIG_DIRECTORY, f), path.join(CONFIG_DIRECTORY, 'legacy', f));
-        fs.rmSync(path.join(CONFIG_DIRECTORY, f))
-      }
-    })
+    // If we are migrating from old cli, then migrate the old config files to the new format
+    if (fs.existsSync(path.join(CONFIG_DIRECTORY, 'keys.json'))) {
+      fs.mkdirSync(path.join(CONFIG_DIRECTORY, 'legacy'), { recursive: true });
+      const filesToMove = ['keys.json', 'config.json', 'chainConfigs.json'];
+      filesToMove.forEach(f => {
+        if (fs.existsSync(path.join(CONFIG_DIRECTORY, f))) {
+          fs.copyFileSync(path.join(CONFIG_DIRECTORY, f), path.join(CONFIG_DIRECTORY, 'legacy', f));
+          fs.rmSync(path.join(CONFIG_DIRECTORY, f))
+        }
+      })
 
-    envConfig.set('envs', envConfig.get('envs').concat('legacy'))
-    envConfig.set('name', 'legacy');
-    createEnv('legacy', envConfig.getProperties());
-    fs.writeFileSync(ROOT_ENV_PATH, JSON.stringify({ default: 'legacy', envs: envConfig.get('envs') }))
+      envConfig.set('envs', envConfig.get('envs').concat('legacy'))
+      defaultEnv = 'legacy';
+      createEnv('legacy', envConfig.getProperties());
+    }
+    envConfig.set('env', defaultEnv);
+    fs.writeFileSync(ROOT_ENV_PATH, JSON.stringify({ default: defaultEnv, envs: envConfig.get('envs') }))
   } catch (error) {
     console.error(error);
   }
